Add files via upload
The Plastic Pathfinder is a model that simulates the transport and accumulation of macroplastics on land. It is a successor of The Trash-Tracker model. The Trash-Tracker simulates the transport and accumulation of macroplastics on land and in the river channel. As the riverine plastic transport modelling was considered too oversimplified by our reviewers, we decided to remove the riverine plastic transport simulations. The Plastic Pathfinder focuses only on the transport and accumulation of macroplastics on land.